added conditional statements
only found that test_write_audio.R needed conditional statements before unlinking directories/files.

+++ R/tests/testthat/test_write_audio.R
@@ -33,15 +33,19 @@ test_that("Audio files are correctly created and removed", {
   expect_true(all(file.exists(file.path(tmp_dir, audio_files))))
 
   # Remove the files created during the test
-  file.remove(file.path(tmp_dir, audio_files))
+  if (file.exists(file.path(tmp_dir, audio_files))) {
+    file.remove(file.path(tmp_dir, audio_files))
+    }
 
   # Remove the test directory
-  unlink(tmp_dir, recursive = TRUE)
+  if (dir.exists(tmp_dir)) {
+    unlink(tmp_dir, recursive = TRUE)
+    }
 })
 
 
 # Unit Test 2: Test for correct df creation (must add an additional column -audio_file_name- for file names with a correct naming format) 
-test_that("Data frame contains correct audio file name format", {  
+test_that("Data frame contains correct audio file name format", {
   # Create a temporary directory on the Desktop for storing files (for test purposes)
   tmp_dir <- file.path(desktop_path, "R_test_temp")
   
@@ -67,8 +71,10 @@ test_that("Data frame contains correct audio file name format", {
   expected_filename <- paste0("TestPrefix_Group", df_test$Group[1], "_Ind", df_test$Individual[1], "_Call", df_test$Call[1], ".wav")
   expect_equal(result_df$audio_file_name[1], expected_filename)
 
-  # Remove the created files
-  unlink(tmp_dir, recursive = TRUE)
+  # Remove the created directory and files
+  if (dir.exists(tmp_dir)) {
+    unlink(tmp_dir, recursive = TRUE)
+    }
 })
 
 # Unit Test 3: Test whether the resulting files contain the correct file extentsion
@@ -97,6 +103,8 @@ test_that("Generated files have the correct .wav extension", {
   audio_files <- result_df$audio_file_name
   expect_true(all(grepl("\\.wav$", audio_files)))
 
-  # Remove the created files
-  unlink(tmp_dir, recursive = TRUE)
+  # Remove the created directory and files
+  if (dir.exists(tmp_dir)) {
+    unlink(tmp_dir, recursive = TRUE)
+    }
 })
